+++ frontend/src/containers/HostEditView/containers/Header/styles.module.scss
@@ -1,0 +1,36 @@
+@import '../../../../styles/colors';
+@import '../../../../styles/fonts';
+@import '../../../../styles/input';
+
+.QuestionEditHeader {
+    width: 100%;
+    height: 100%;
+    padding: 3.5vh 1%;
+    box-sizing: border-box;
+    background-image: $grad-purple-light;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    align-content: center;
+    align-items: center;
+}
+
+.QuestionEditHeader * {
+    height: 100%;
+}
+
+.titleInput {
+    @extend %text;
+    width: 50%;
+    font-family: $font-text;
+    font-size: $font-md;
+    background-color: #FFFFFF47;
+}
+
+.titleInput:focus {
+    background-color: $white;
+}
+
+.buttons {
+    width: 44%;
+}
