+++ frontend/src/containers/HostEditView/containers/QuestionEditor/components/Points/styles.module.scss
@@ -1,0 +1,28 @@
+@import '../../../../../../styles/colors';
+@import '../../../../../../styles/fonts';
+@import '../../../../../../styles/input';
+
+.Points {
+  width: 150px;
+  height: 200px;
+}
+
+.choices {
+  height: 150px;
+  display: flex;
+  flex-direction: column;
+}
+
+.choices button.choice {
+  @extend %button;
+  height: 50px;
+  border-radius: 100%;
+  background-color: $magenta-dark;
+  color: $white;
+  font-family: $font-text;
+  font-size: $font-sm;
+}
+
+.choices button.choice.selected {
+  background-color: $pink;
+}
