+++ frontend/src/containers/HostEditView/containers/QuestionEditor/containers/Overlay/styles.module.scss
@@ -1,0 +1,10 @@
+@import '../../../../../../styles/scroller';
+
+.container {
+  margin: 16px 8px 16px 8px;
+  display: flex;
+  flex-wrap: wrap;
+  // justify-content: space-evenly;
+  // align-content: center;
+  // align-items: center;
+}
