+++ frontend/src/containers/HostEditView/containers/QuestionEditor/styles.module.scss
@@ -1,0 +1,58 @@
+@import '../../../../styles/colors';
+@import '../../../../styles/fonts';
+@import '../../../../styles/scroller';
+@import '../../../../styles/input';
+
+.QuestionEditor {
+  // margin: auto 0;
+  padding: 0 32px;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+}
+
+.QuestionEditor > div {
+  margin-bottom: 2vh;
+}
+
+.top {
+  margin-top: 16px;
+  display: flex;
+}
+
+.mid {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+}
+
+.question {
+  @extend %textarea;
+  margin-right: 24px;
+  min-width: 500px;
+  min-height: 200px;
+  flex: 1;
+  font-family: $font-text;
+  font-size: $font-md;
+  background-color: $neon;
+}
+
+.question:focus {
+  background-color: $white;
+}
+
+.settings {
+  flex: 1;
+  display: flex;
+  justify-content: space-evenly;
+  align-content: center;
+  align-items: center;
+}
+
+.answers {
+  flex: 1;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+}
