+++ frontend/src/containers/HostEditView/containers/Sidebar/containers/QuestionList/components/QuestionListCard/styles.module.scss
@@ -1,0 +1,50 @@
+@import '../../../../../../../../styles/colors';
+@import '../../../../../../../../styles/input';
+
+$card-width: 216px;
+$card-height: 148px;
+
+.QuestionListCard {
+  width: $card-width;
+  height: $card-height;
+  padding: 10px 10px;
+  border-radius: 17px;
+  color: $black;
+  background-color: #fff;
+  font-size: 1.2rem;
+  display: flex;
+  justify-content: flex-start;
+  justify-items: center;
+  align-items: center;
+}
+
+.icons {
+  margin-right: 8px;
+  width: 40px;
+  height: 80%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  justify-items: center;
+  align-items: center;
+}
+
+.question {
+  @extend %button;
+  padding: 0;
+  max-width: 160px;
+  flex: 1;
+  height: 100%;
+  font-size: 1.2rem;
+  background-color: transparent;
+  white-space: pre-line;
+  overflow: hidden;
+  display: -webkit-box;
+  text-overflow: ellipsis;
+  -webkit-line-clamp: 5;
+  -webkit-box-orient: vertical;
+}
+
+.question:hover {
+  background-color: #00000017;
+}
