+++ frontend/src/containers/HostEditView/containers/Sidebar/containers/QuestionList/styles.module.scss
@@ -1,0 +1,15 @@
+@import '../../../../../../styles/colors';
+@import '../../../../../../styles/scroller';
+
+.QuestionList {
+  padding-top: 8px;
+  width: 100%;
+  height: 62vh;
+  padding-left: $scroll-width;
+  overflow-y: scroll;
+  background-color: #00000040;
+}
+
+.QuestionList > * {
+  margin: 0 8px 8px 8px;
+}
