+++ frontend/src/containers/HostEditView/containers/Sidebar/styles.module.scss
@@ -1,0 +1,11 @@
+@import '../../../../styles/colors';
+
+.QuestionSidebar {
+  width: 100%;
+  height: 100%;
+  display: grid;
+  grid-template-columns: 100%;
+  grid-template-rows: 80% 20%;
+  justify-content: center;
+  align-content: center;
+}
