--- frontend/src/styles/AnswerCard.module.scss
@@ -1,24 +1,0 @@
-@import 'answer';
-
-.AnswerCard {
-  margin-bottom: 24px;
-  margin: auto 0;
-  padding: 0 24px;
-  width: 500px;
-  height: 150px;
-  border-radius: 17px;
-  display: flex;
-  justify-content: flex-start;
-  align-content: center;
-  align-items: center;
-}
-
-.shape {
-  margin-right: 24px;
-}
-
-.label {
-  width: auto;
-  font-size: 2rem;
-  font-weight: bold;
-}
