--- frontend/src/styles/AnswerEditCard.module.scss
@@ -1,32 +1,0 @@
-@import 'input';
-@import 'answer';
-
-.card {
-  margin-bottom: 24px;
-  margin: auto 0;
-  padding: 24px;
-  width: 500px;
-  height: 150px;
-  border-radius: 17px;
-  display: flex;
-  justify-content: space-between;
-  align-content: center;
-  align-items: center;
-}
-
-.shape {
-  margin-right: 24px;
-}
-
-.text {
-  @extend %textarea;
-  margin-right: 24px;
-  height: 100%;
-  flex: 1;
-  font-size: 2rem;
-  background-color: #00000017;
-}
-
-.text:focus {
-  background-color: $white;
-}
