--- frontend/src/styles/HostEditView.module.scss
@@ -1,33 +1,0 @@
-@import 'scroller';
-@import 'colors';
-
-.HostEditView {
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  position: absolute;
-  display: grid;
-  grid-template-columns: 100%;
-  grid-template-rows: 18vh 82vh;
-  justify-content: center;
-  align-content: center;
-  justify-items: flex-start;
-  align-items: center;
-}
-
-.header {
-  background-color: $grad-purple-light;
-}
-
-.content {
-  width: 100%;
-  height: 100%;
-  display: grid;
-  grid-template-columns: 262px auto;
-  grid-template-rows: auto;
-}
-
-.editor {
-  margin: 0 auto;
-}
