--- frontend/src/styles/MoreQuestions.module.scss
@@ -1,11 +1,0 @@
-@import 'scroller';
-@import 'MoreQuestionsCard.module.scss';
-
-.container {
-  margin: 16px 8px 16px 8px;
-  display: flex;
-  flex-wrap: wrap;
-  // justify-content: space-evenly;
-  // align-content: center;
-  // align-items: center;
-}
