--- frontend/src/styles/Points.module.scss
@@ -1,28 +1,0 @@
-@import 'colors';
-@import 'fonts';
-@import 'input';
-
-.Points {
-  width: 150px;
-  height: 200px;
-}
-
-.choices {
-  height: 150px;
-  display: flex;
-  flex-direction: column;
-}
-
-.choices button.choice {
-  @extend %button;
-  height: 50px;
-  border-radius: 100%;
-  background-color: $magenta-dark;
-  color: $white;
-  font-family: $font-text;
-  font-size: $font-sm;
-}
-
-.choices button.choice.selected {
-  background-color: $pink;
-}
