--- frontend/src/styles/QuestionList.module.scss
@@ -1,15 +1,0 @@
-@import 'colors';
-@import 'scroller';
-
-.QuestionList {
-  padding-top: 8px;
-  width: 100%;
-  height: 62vh;
-  padding-left: $scroll-width;
-  overflow-y: scroll;
-  background-color: #00000040;
-}
-
-.QuestionList > * {
-  margin: 0 8px 8px 8px;
-}
