--- frontend/src/styles/QuestionSidebar.module.scss
@@ -1,11 +1,0 @@
-@import 'colors';
-
-.QuestionSidebar {
-  width: 100%;
-  height: 100%;
-  display: grid;
-  grid-template-columns: 100%;
-  grid-template-rows: 80% 20%;
-  justify-content: center;
-  align-content: center;
-}
