--- frontend/src/styles/Switch.module.scss
@@ -1,44 +1,0 @@
-@import 'fonts';
-@import 'colors';
-
-.Switch {
-  width: 48px;
-  height: 96px;
-  background-color: $black;
-}
-
-.container {
-  width: min-content;
-  height: min-content;
-  background-color: $black;
-  display: flex;
-  flex-direction: column;
-}
-
-.container button.top,
-.container button.bottom {
-  margin: 4px;
-  width: 48px;
-  height: 48px;
-  padding: none;
-  border: none;
-  outline: none;
-  font-size: $font-sm;
-  background-color: $purple-very-dark;
-  display: flex;
-  justify-content: center;
-  align-content: center;
-  align-items: center;
-}
-
-.container button.bottom {
-  margin-top: 0;
-}
-
-.container button.top.selected {
-  background-color: $green;
-}
-
-.container button.bottom.selected {
-  background-color: $red;
-}
